refactor #95: OrderErrorCode 경로 변경
